+++ src/windows/volume.h
@@ -19,6 +19,7 @@
 
 #include <windows.h>
 #include <Rpc.h>
+#include <Shlobj.h>
 #include <vector>
 #include "src/mountpoint.h"
 #include "src/windows/com.h"
@@ -37,6 +38,8 @@ enum class Type {
   RAM_DISK
 };
 
+HRESULT GetSystemVolume(wchar_t *out);
+HRESULT GetAvailableVolumes(std::vector<wchar_t> *const output);
 HANDLE OpenHandle(const wchar_t letter, DWORD flags);
 HRESULT GetDeviceNumber(const wchar_t letter, ULONG *out);
 HRESULT IsDiskWritable(const wchar_t letter, BOOL *out);